+++ src/components/atoms/ButtonSample.tsx
@@ -1,4 +1,4 @@
-import ThemeContext from "@/app/use-context-sample/theme-context";
+import ThemeContext from "@/app/use-context-sample/themeContext";
 import { useContext } from "react";
 
 export default function ButtonSample() {
